# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Music Media Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /music/tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List all tracks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 100
        - name: search
          in: query
          schema:
            type: string
        - name: albumId
          in: query
          schema:
            type: integer
        - name: artistId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackListResponse"
  /music/tracks/{id}:
    get:
      operationId: getTrack
      tags: [tracks]
      summary: Get track details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /music/stream/{id}:
    get:
      operationId: streamTrack
      tags: [tracks]
      summary: Stream audio file for a track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audio stream
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "206":
          description: Partial audio content (range request)
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
          nullable: true
        artistName:
          type: string
          nullable: true
        albumId:
          type: integer
          nullable: true
        albumTitle:
          type: string
          nullable: true
        trackNumber:
          type: integer
          nullable: true
        discNumber:
          type: integer
          nullable: true
        durationSeconds:
          type: number
          nullable: true
        genre:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        filePath:
          type: string
        mimeType:
          type: string
        hasArt:
          type: boolean
      required:
        - id
        - title
        - filePath
        - mimeType
        - hasArt
    TrackListResponse:
      type: object
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - tracks
        - total
        - page
        - pageSize
...